+++ main.go
@@ -12,7 +12,7 @@ import (
 	service "github.com/arangodb-helper/testagent/service"
 	arangodb "github.com/arangodb-helper/testagent/service/cluster/arangodb"
 	"github.com/arangodb-helper/testagent/service/test"
-	"github.com/arangodb-helper/testagent/tests/replication2"
+	complex "github.com/arangodb-helper/testagent/tests/complex"
 	"github.com/arangodb-helper/testagent/tests/simple"
 	logging "github.com/op/go-logging"
 	"github.com/pkg/errors"
@@ -41,9 +41,9 @@ var (
 		service.ServiceConfig
 		arangodb.ArangodbConfig
 		simple.SimpleConfig
-		replication2.Replication2Config
-		replication2.DocColConfig
-		replication2.GraphTestConf
+		complex.ComplextTestConfig
+		complex.DocColConfig
+		complex.GraphTestConf
 		logLevel string
 	}
 	maskAny = errors.WithStack
@@ -74,19 +74,19 @@ func init() {
 	f.IntVar(&appFlags.SimpleConfig.MaxCollections, "simple-max-collections", 10, "Upper limit to the number of collections created in simple test")
 	f.DurationVar(&appFlags.SimpleConfig.OperationTimeout, "simple-operation-timeout", defaultOperationTimeout, "Timeout per database operation")
 	f.DurationVar(&appFlags.SimpleConfig.RetryTimeout, "simple-retry-timeout", defaultRetryTimeout, "How long are tests retried before giving up")
-	f.Int64Var(&appFlags.GraphTestConf.MaxVertices, "graph-max-vertices", 100000, "Upper limit to the number of vertices created in community graph test")
-	f.IntVar(&appFlags.GraphTestConf.VertexSize, "graph-vertex-size", 512, "The size of the payload field in bytes in all vertices in community graph test")
-	f.IntVar(&appFlags.GraphTestConf.EdgeSize, "graph-edge-size", 128, "The size of the payload field in bytes in all vertices in community graph test")
-	f.IntVar(&appFlags.GraphTestConf.TraversalOperationsPerCycle, "graph-traversal-ops", 1000, "How many traversal operations to perform in one test cycle")
+	f.Int64Var(&appFlags.GraphTestConf.MaxVertices, "graph-max-vertices", 100000, "Upper limit to the number of vertices created in all graph tests")
+	f.IntVar(&appFlags.GraphTestConf.VertexSize, "graph-vertex-size", 512, "The size of the payload field in bytes in all vertices in all graph tests")
+	f.IntVar(&appFlags.GraphTestConf.EdgeSize, "graph-edge-size", 128, "The size of the payload field in bytes in all vertices in all graph tests")
+	f.IntVar(&appFlags.GraphTestConf.TraversalOperationsPerCycle, "graph-traversal-ops", 1000, "How many traversal operations to perform in one test cycle all graph tests")
 	f.IntVar(&appFlags.GraphTestConf.BatchSize, "graph-batch-size", 1000, "Batch size for creating documents in graph tests")
-	f.IntVar(&appFlags.DocColConfig.MaxDocuments, "replication2-max-documents", 1000000, "Upper limit to the number of documents created in replication2 test")
-	f.IntVar(&appFlags.DocColConfig.BatchSize, "replication2-batch-size", 1000, "Batch size for creating documents in bulk mode in replication2 tests")
-	f.IntVar(&appFlags.DocColConfig.DocumentSize, "replication2-document-size", 1024, "The size of the payload field in bytes in all documents in replication2 test")
-	f.IntVar(&appFlags.DocColConfig.MaxUpdates, "replication2-max-updates", 10, "Number of update operations to be performed on each document, before dropping collection.")
-	f.IntVar(&appFlags.Replication2Config.NumberOfShards, "replication2-shards", 10, "Number of shards in all replication2 tests")
-	f.IntVar(&appFlags.Replication2Config.ReplicationFactor, "replication2-replicationFactor", 3, "Replication factor in all replication2 tests")
-	f.DurationVar(&appFlags.Replication2Config.OperationTimeout, "replication2-operation-timeout", defaultOperationTimeout, "Timeout per database operation in all replication2 tests")
-	f.DurationVar(&appFlags.Replication2Config.RetryTimeout, "replication2-retry-timeout", defaultRetryTimeout, "How long are tests retried before giving up in all replication2 tests")
+	f.IntVar(&appFlags.DocColConfig.MaxDocuments, "complex-max-documents", 1000000, "Upper limit to the number of documents created in tests from the complex suite")
+	f.IntVar(&appFlags.DocColConfig.BatchSize, "complex-batch-size", 1000, "Batch size for creating documents in bulk mode in in tests from the complex suite")
+	f.IntVar(&appFlags.DocColConfig.DocumentSize, "complex-document-size", 1024, "The size of the payload field in bytes in regular documents in tests from the complex suite")
+	f.IntVar(&appFlags.DocColConfig.MaxUpdates, "complex-max-updates", 10, "Number of update operations to be performed on each document, before dropping collection.")
+	f.IntVar(&appFlags.ComplextTestConfig.NumberOfShards, "complex-shards", 10, "Number of shards in in tests from the complex suite")
+	f.IntVar(&appFlags.ComplextTestConfig.ReplicationFactor, "complex-replicationFactor", 3, "Replication factor in tests from the complex suite")
+	f.DurationVar(&appFlags.ComplextTestConfig.OperationTimeout, "complex-operation-timeout", defaultOperationTimeout, "Timeout per database operation in tests from the complex suite")
+	f.DurationVar(&appFlags.ComplextTestConfig.RetryTimeout, "complex-retry-timeout", defaultRetryTimeout, "How long are tests retried before giving up in tests from the complex suite")
 }
 
 // handleSignal listens for termination signals and stops this process onup termination.
@@ -165,11 +165,11 @@ func cmdMainRun(cmd *cobra.Command, args []string) {
 
 	// Create tests
 	tests := []test.TestScript{
-		replication2.NewDocColTest(log, appFlags.ReportDir, appFlags.Replication2Config, appFlags.DocColConfig),
-		replication2.NewOneShardTest(log, appFlags.ReportDir, appFlags.Replication2Config, appFlags.DocColConfig),
-		replication2.NewComGraphTest(log, appFlags.ReportDir, appFlags.Replication2Config, appFlags.GraphTestConf),
-		replication2.NewSmartGraphTest(log, appFlags.ReportDir, appFlags.Replication2Config, appFlags.GraphTestConf),
-		replication2.NewEnterpriseGraphTest(log, appFlags.ReportDir, appFlags.Replication2Config, appFlags.GraphTestConf),
+		complex.NewDocColTest(log, appFlags.ReportDir, appFlags.ComplextTestConfig, appFlags.DocColConfig),
+		complex.NewOneShardTest(log, appFlags.ReportDir, appFlags.ComplextTestConfig, appFlags.DocColConfig),
+		complex.NewComGraphTest(log, appFlags.ReportDir, appFlags.ComplextTestConfig, appFlags.GraphTestConf),
+		complex.NewSmartGraphTest(log, appFlags.ReportDir, appFlags.ComplextTestConfig, appFlags.GraphTestConf),
+		complex.NewEnterpriseGraphTest(log, appFlags.ReportDir, appFlags.ComplextTestConfig, appFlags.GraphTestConf),
 		simple.NewSimpleTest(log, appFlags.ReportDir, appFlags.SimpleConfig),
 	}
