+++ tests/complex/database.go
@@ -1,4 +1,4 @@
-package replication2
+package complex
 
 import (
 	"fmt"
@@ -7,28 +7,30 @@ import (
 	"github.com/arangodb-helper/testagent/service/test"
 )
 
-// createCollection creates a new collection.
+func (t *ComplextTest) createOneShardDatabase(databaseName string) error {
+	return t.createDatabase(databaseName, "single", t.ReplicationFactor, t.ReplicationFactor-1)
+}
+
+// createDatabase creates a new database.
 // The operation is expected to succeed.
-func (t *Replication2Test) createCollection(collectionName string, edge bool) error {
-	var colType int
-	var colTypeName string
-	if edge {
-		colType = 3
-		colTypeName = "edge"
-	} else {
-		colType = 2
-		colTypeName = "document"
-	}
-	opts := struct {
-		Name              string `json:"name"`
-		NumberOfShards    int    `json:"numberOfShards"`
-		ReplicationFactor int    `json:"replicationFactor"`
-		Type              int    `json:"type"`
-	}{
-		Name:              collectionName,
-		NumberOfShards:    t.NumberOfShards,
-		ReplicationFactor: t.ReplicationFactor,
-		Type:              colType,
+func (t *ComplextTest) createDatabase(databaseName string, sharding string, replicationFactor int, writeConcern int) error {
+	body := struct {
+		Name    string `json:"name"`
+		Options struct {
+			Sharding          string `json:"sharding"`
+			ReplicationFactor int    `json:"replicationFactor"`
+			WriteConcern      int    `json:"writeConcern"`
+		} `json:"options"`
+	}{Name: databaseName,
+		Options: struct {
+			Sharding          string `json:"sharding"`
+			ReplicationFactor int    `json:"replicationFactor"`
+			WriteConcern      int    `json:"writeConcern"`
+		}{
+			Sharding:          sharding,
+			ReplicationFactor: replicationFactor,
+			WriteConcern:      writeConcern,
+		},
 	}
 	//operationTimeout, retryTimeout := t.OperationTimeout, t.RetryTimeout
 	// For now, we increase the timeout to 5 minutes, since the cluster-internal
@@ -51,10 +53,9 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 		shouldNotExist := false
 		shouldExist := false
 
-		t.log.Infof("Creating (%d) collection '%s' of type '%s' with numberOfShards=%d, replicationFactor=%d...",
-			i, collectionName, colTypeName, t.NumberOfShards, t.ReplicationFactor)
+		t.log.Infof("Creating database '%s'...", databaseName)
 		resp, err := t.client.Post(
-			"/_api/collection", nil, nil, opts, "", nil, []int{0, 1, 200, 409, 500, 503},
+			"/_api/database", nil, nil, body, "", nil, []int{0, 1, 201, 409, 500, 503},
 			[]int{400, 404, 307}, operationTimeout, 1)
 		t.log.Infof("... got http %d - arangodb %d via %s",
 			resp[0].StatusCode, resp[0].Error_.ErrorNum, resp[0].CoordinatorURL)
@@ -74,7 +75,7 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 		//         else : failure
 
 		if err[0] == nil {
-			if resp[0].StatusCode == 200 {
+			if resp[0].StatusCode == 201 {
 				success = true
 			} else {
 				if resp[0].StatusCode == 1 || resp[0].StatusCode == 500 { // connection refused or not created
@@ -84,9 +85,9 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 				} else if resp[0].StatusCode == 409 {
 					if i == 1 {
 						// This is a failure
-						t.createCollectionCounter.failed++
-						t.reportFailure(test.NewFailure(t.Name(), "Failed to create collection '%s': got 409 on first attempt", collectionName))
-						return maskAny(fmt.Errorf("Failed to create collection '%s': got 409 on first attempt", collectionName))
+						t.createDatabaseCounter.failed++
+						t.reportFailure(test.NewFailure(t.Name(), "Failed to create database '%s': got 409 on first attempt", databaseName))
+						return maskAny(fmt.Errorf("Failed to create database '%s': got 409 on first attempt", databaseName))
 					} else {
 						shouldExist = true
 					}
@@ -95,34 +96,34 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 			}
 		} else {
 			// This is a failure
-			t.createCollectionCounter.failed++
-			t.reportFailure(test.NewFailure(t.Name(), "Failed to create collection '%s': %v", collectionName, err[0]))
+			t.createDatabaseCounter.failed++
+			t.reportFailure(test.NewFailure(t.Name(), "Failed to create database '%s': %v", databaseName, err[0]))
 			return maskAny(err[0])
 		}
 
 		if checkRetry {
 
-			t.log.Infof("Checking existence of collection '%s' ...", collectionName)
-			exists, checkErr := t.collectionExists(collectionName)
+			t.log.Infof("Checking existence of database '%s' ...", databaseName)
+			exists, checkErr := t.databaseExists(databaseName)
 			t.log.Infof("... got result %v and error %v", exists, checkErr)
 
 			if checkErr == nil {
 				if exists {
 					if shouldNotExist {
 						// This is a failure
-						t.createCollectionCounter.failed++
+						t.createDatabaseCounter.failed++
 						t.reportFailure(test.NewFailure(t.Name(),
-							"Failed to create collection '%s' rechecked and failed existence", collectionName))
-						return maskAny(fmt.Errorf("Failed to create collection '%s' rechecked and failed existence", collectionName))
+							"Failed to create database '%s' rechecked and failed existence", databaseName))
+						return maskAny(fmt.Errorf("Failed to create database '%s' rechecked and failed existence", databaseName))
 					}
 					success = true
 				} else {
 					if shouldExist {
 						// This is a failure
-						t.createCollectionCounter.failed++
+						t.createDatabaseCounter.failed++
 						t.reportFailure(test.NewFailure(t.Name(),
-							"Failed to create collection '%s' rechecked and failed existence", collectionName))
-						return maskAny(fmt.Errorf("Failed to create collection '%s' rechecked and failed existence", collectionName))
+							"Failed to create database '%s' rechecked and failed existence", databaseName))
+						return maskAny(fmt.Errorf("Failed to create database '%s' rechecked and failed existence", databaseName))
 					}
 				}
 			} else {
@@ -131,10 +132,9 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 		}
 
 		if success {
-			t.createCollectionCounter.succeeded++
+			t.createDatabaseCounter.succeeded++
 			t.log.Infof(
-				"Creating collection '%s' of type '%s' with numberOfShards=%d, replicationFactor=%d succeeded",
-				collectionName, colTypeName, t.NumberOfShards, t.ReplicationFactor)
+				"Creating database '%s' succeeded", databaseName)
 			return nil
 		}
 
@@ -147,19 +147,19 @@ func (t *Replication2Test) createCollection(collectionName string, edge bool) er
 
 	// Overall timeout :(
 	t.reportFailure(
-		test.NewFailure(t.Name(), "Timed out while trying to create (%d) collection %s.", i, collectionName))
-	return maskAny(fmt.Errorf("Timed out while trying to create (%d) collection %s.", i, collectionName))
+		test.NewFailure(t.Name(), "Timed out while trying to create (%d) database %s.", i, databaseName))
+	return maskAny(fmt.Errorf("Timed out while trying to create (%d) database %s.", i, databaseName))
 
 }
 
-// dropCollection remove an existing collection.
+// dropDatabase removes an existing database.
 // The operation is expected to succeed.
-func (t *Replication2Test) dropCollection(collectionName string) error {
+func (t *ComplextTest) dropDatabase(databaseName string) error {
 
 	operationTimeout := t.OperationTimeout
 	testTimeout := time.Now().Add(t.OperationTimeout * 5)
 
-	url := fmt.Sprintf("/_api/collection/%s", collectionName)
+	url := fmt.Sprintf("/_api/database/%s", databaseName)
 	backoff := time.Millisecond * 250
 	i := 0
 
@@ -171,26 +171,26 @@ func (t *Replication2Test) dropCollection(collectionName string) error {
 			break
 		}
 
-		t.log.Infof("Removing (%d) collection '%s'...", i, collectionName)
+		t.log.Infof("Removing (%d) database '%s'...", i, databaseName)
 		resp, err := t.client.Delete(
 			url, nil, nil, []int{0, 1, 200, 404, 500, 503}, []int{400, 409, 307}, operationTimeout, 1)
 		t.log.Infof("... got http %d - arangodb %d", resp[0].StatusCode, resp[0].Error_.ErrorNum)
 
 		if err[0] != nil {
 			// This is a failure
-			t.dropCollectionCounter.failed++
-			t.reportFailure(test.NewFailure(t.Name(), "Failed to drop collection '%s': %v", collectionName, err[0]))
+			t.dropDatabaseCounter.failed++
+			t.reportFailure(test.NewFailure(t.Name(), "Failed to drop database '%s': %v", databaseName, err[0]))
 			return maskAny(err[0])
 		} else if resp[0].StatusCode == 404 {
-			// Collection not found.
+			// Database not found.
 			// This can happen if the first attempt timed out, but did actually succeed.
 			// So we accept this if there are multiple attempts.
 			if i == 1 { // this is a failure in first run
 				// Not enough attempts, this is a failure
-				t.dropCollectionCounter.failed++
+				t.dropDatabaseCounter.failed++
 				t.reportFailure(
-					test.NewFailure(t.Name(), "Failed to drop collection '%s': got 404 after only 1 attempt", collectionName))
-				return maskAny(fmt.Errorf("Failed to drop collection '%s': got 404 after only 1 attempt", collectionName))
+					test.NewFailure(t.Name(), "Failed to drop database '%s': got 404 after only 1 attempt", databaseName))
+				return maskAny(fmt.Errorf("Failed to drop database '%s': got 404 after only 1 attempt", databaseName))
 			} else {
 				success = true
 			}
@@ -199,8 +199,8 @@ func (t *Replication2Test) dropCollection(collectionName string) error {
 		}
 
 		if success {
-			t.dropCollectionCounter.succeeded++
-			t.log.Infof("Droping collection '%s' succeeded", collectionName)
+			t.dropDatabaseCounter.succeeded++
+			t.log.Infof("Droping database '%s' succeeded", databaseName)
 			return nil
 		}
 
@@ -211,20 +211,26 @@ func (t *Replication2Test) dropCollection(collectionName string) error {
 
 	}
 
-	t.dropCollectionCounter.failed++
-	t.reportFailure(test.NewFailure(t.Name(), "Timed out (%d) while droping collection '%s'", i, collectionName))
-	return maskAny(fmt.Errorf("Timed out (%d) while droping collection '%s'", i, collectionName))
+	t.dropDatabaseCounter.failed++
+	t.reportFailure(test.NewFailure(t.Name(), "Timed out (%d) while droping database '%s'", i, databaseName))
+	return maskAny(fmt.Errorf("Timed out (%d) while droping database '%s'", i, databaseName))
 
 }
 
-func (t *Replication2Test) collectionExists(collectionName string) (bool, error) {
+func (t *ComplextTest) databaseExists(databaseName string) (bool, error) {
+
+	type DatabasesResponse struct {
+		Error  bool
+		Code   string
+		Result []string
+	}
 
 	operationTimeout := time.Duration(ReadTimeout) * time.Second
 	timeout := time.Now().Add(operationTimeout)
 
 	i := 0
 	backoff := time.Millisecond * 250
-	url := fmt.Sprintf("/_api/collection/%s", collectionName)
+	url := fmt.Sprintf("/_api/database")
 
 	for {
 
@@ -233,19 +239,25 @@ func (t *Replication2Test) collectionExists(collectionName string) (bool, error)
 			break
 		}
 
-		t.log.Infof("Checking (%d) collection '%s'...", i, collectionName)
+		t.log.Infof("Checking (%d) database '%s'...", i, databaseName)
+		var response DatabasesResponse
 		resp, err := t.client.Get(
-			url, nil, nil, nil, []int{0, 1, 200, 404, 503}, []int{400, 409, 307}, operationTimeout, 1)
+			url, nil, nil, response, []int{0, 1, 200, 404, 503}, []int{400, 409, 307}, operationTimeout, 1)
 		t.log.Infof("... got http %d - arangodb %d", resp[0].StatusCode, resp[0].Error_.ErrorNum)
 
 		if err[0] != nil {
 			// This is a failure
-			t.log.Infof("Failed checking for collection '%s': %v", collectionName, err[0])
+			t.log.Infof("Failed checking for database '%s': %v", databaseName, err[0])
 			return false, maskAny(err[0])
 		} else if resp[0].StatusCode == 404 {
 			return false, nil
 		} else if resp[0].StatusCode == 200 {
-			return true, nil
+			for k := 0; k < len(response.Result); k++ {
+				if response.Result[k] == databaseName {
+					return true, nil
+				}
+			}
+			return false, nil
 		}
 
 		// 0, 1, 503 retry
@@ -257,7 +269,7 @@ func (t *Replication2Test) collectionExists(collectionName string) (bool, error)
 	}
 
 	// This is a failure
-	out := fmt.Errorf("Timed out checking for collection '%s'", collectionName)
+	out := fmt.Errorf("Timed out checking for database '%s'", databaseName)
 	t.log.Error(out)
 	return false, maskAny(out)
