+++ tests/complex/error.go
@@ -1,4 +1,4 @@
-package replication2
+package complex
 
 import (
 	"github.com/pkg/errors"
