+++ tests/complex/util.go
@@ -1,4 +1,4 @@
-package replication2
+package complex
 
 import "math/rand"
